requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
pydantic
sentence-transformers